feat: seperate networkModule, addObserver for notification
